requirements.txt
Django==3.1.7
djangorestframework==3.12.2
# Dev dependencies
pytest-django==4.1.0
mock==4.0.3
